[Refactor] Config SQLite migration
